+++ pdf/lib/src/widgets/text.dart
@@ -773,7 +773,11 @@ class RichText extends Widget with SpanningWidget {
 
       for (var index = 0; index < text.length; index++) {
         final rune = text[index];
-        if (rune == 0x0a) {
+        const spaces = {
+          0x0a, 0x09, 0x00A0, 0x1680, 0x2000, 0x2001, 0x2002, 0x2003, 0x2004, //
+          0x2005, 0x2006, 0x2007, 0x2008, 0x2009, 0x200A, 0x202F, 0x205F, 0x3000
+        };
+        if (spaces.contains(rune)) {
           continue;
         }
 
@@ -905,6 +909,7 @@ class RichText extends Widget with SpanningWidget {
 
           for (var line = 0; line < spanLines.length; line++) {
             final words = spanLines[line].split(RegExp(r'\s'));
+            print(words);
             for (var index = 0; index < words.length; index++) {
               final word = words[index];